Save before Adding new Analysis
- Next few commits will involve adding ANOVA analyses and better statistical testing.
- This commit has a few things changes made during the preceeding weeks, notably:
preprocessLogFileMonkeyLogic has a bug fix related to not having frame labels, and a proper cross checking of independently constructed and saved stimulus lists has been implemented.
excludeTrialsMonkeyLogic has a new block of code which enables the exclusion of correct trials following a failed trial. It can be set to 0 to have no effect.
runAnalyses has changes in "imageEyeMap" based on using the "rat" function to downsample instead of previous technique. Function isn't complete yet, but will be in the coming weeks.
Added brief line of code at the end of my Parambuilder file so as to throw an error if the analogFile doesn't exist (instead of needing to see it reach much further).