stimCatANOVA completed
Finished "stimCatANOVA" function, which creates an structure for every unit in every epoch describes the results of a one-way ANOVA, performed either between all event IDs (if target = 0), or between all events with a label matching the string in "target" and those without this label. [Labels come from "group" structure, which will change in the event of certain analyses which create different groups). Proper way to do this would also likely involve 2-way ANOVA, with Epoch and Label as factors.