behaviorsummaryPhyzzy upgrade
Changed the management of the excel record (if detectable) for files. Briefly, old behavior had it that if the userVars had clipDuration, it acted as a binary indicator for whether all other variables are present, and overwrote them with the excel doc. Now, the presence of the correctly dated line on the excel decides whether variables will be pulled from it, and only present variables on excel will overwrite those in the data's structure.
- Got rid of recording threshold since it didn't seem useful, implemented reward frequency, to account for future trials where the reward frequency won't be 1.
- Removed some subplot numbers, which are redundant if position info is given.
-