Merge plotzzy
Wrote a brief set of scripts which allow for summary documents to be generated. As to not complicate life with directory managmenet issues, I've opted to place those scripts into this main directory.

Added brief bit of code at the end of processRunBatch to allow for this code to be run at the end of a batch run.

updated export+fig to latest commit on github.