clusterFix dual filter, processRunBatch changes, runAnalyses
ClusterFix now internally processing eye signal twice - Once with a 30 Hz low pass filter for export to other functions, and then with a 2nd filter (set in analysisParam) for internal processing. I may need to tune this function to avoid detecting microsaccades if attempts to do so outside the function fail.
- made changes to how saccade down sampling happens (converted to interp1 like the rest).
- added option to add frame number to the lower left of the video for eyeStimOverlay.
- added simple saccade filtration based on mean distance (may need to revise).