figStruct, spikePupilCorr close to completion, pupilDilation NaNs for blinks.
figStruct - a struct generated in analysisParam which combines fig relevant parameters (many of the switch near the top of the default analysisParam + some new ones) and packages 'channelUnitNames' for the sake of appending to titles.
runAnalyses - improved argument handlings where parameter structs don't undergo revision outside of the switches which use them (allows for a bit more transparency and less dependency issues).
spikePupilCorr - updated to produce figures with the right titles, as well as a "Per Stim" style image, with all the units plotted, and a "Per Unit" style image. One form of the "Per Unit" style image is in need of revision, as it is deeply redundant in some ways. 3 switches sit inside this function - 1 to exclude pre and post stim period, a 'plotType' to decide between the two mentioned plots, a 2nd 'figureType' which decides between the 2 'plotType = 2" plots. (Bit messy...).
pupilDilation now places NaNs in the place of blinks to give a more accurate picture of values known.